use default solidus implementation for non-assembly products
