+++ app/models/spree/stock/availability_validator_decorator.rb
@@ -1,0 +1,3 @@
+Spree::Stock::AvailabilityValidator.class_eval do
+  prepend Spree::Stock::AssemblyAvailabilityValidator
+end
